Add sample and compiler project
