+++ mc.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+dotnet run --project ./mc/mc.csproj -- "$@"
