+++ samples/hello/helloworld.ms
@@ -1,0 +1,3 @@
+print("What is your name?")
+let name = input()
+print("Hello " + name + "!")
